+++ spec/image_optim/config_spec.rb
@@ -45,24 +45,24 @@ describe ImageOptim::Config do
     end
   end
 
-  describe :always_replace do
+  describe :skip_bigger do
     before do
       allow(IOConfig).to receive(:read_options).and_return({})
     end
 
     it 'is true by default' do
       config = IOConfig.new({})
-      expect(config.always_replace).to be true
+      expect(config.skip_bigger).to be false
     end
 
     it 'is false when false' do
-      config = IOConfig.new(:always_replace => false)
-      expect(config.always_replace).to be false
+      config = IOConfig.new(:skip_bigger => true)
+      expect(config.skip_bigger).to be true
     end
 
     it 'is false when false' do
-      config = IOConfig.new(:always_replace => true)
-      expect(config.always_replace).to be true
+      config = IOConfig.new(:skip_bigger => false)
+      expect(config.skip_bigger).to be false
     end
   end